Change packet structure
Do not update offset and header metadata for the moment. Two reasons
(a) We are not using this at the moment.
(b) We end up writing to the packet more frequently than desirable.

Writing out offset is moved behind a feature flag rather than being entirely disabled, allowing us to go back if
desirable.