Hopefully speed up Travis by using Docker (#27)
Run build in a container, that is already mostly set up (e.g., no building Cargo). The main problem is that this is currently not very generic (e.g., the builds are done as `root`, making `dist_clean` and other things a bit painful to run). But it does speed up the Travis jobs significantly.