test tcp window creation_test fails, cannot alloc > 1024MB (#62)
dpdk requires 1024MB Hugepages to run. But when the host
has allocated exactly that amount, tcp_window.rs creation_test
fails as when i == 1073741824, ReorderedBuffer::new(i + 1) asks
for more bytes than we have in Hugepages.

Change so the max i is /2. new(i + 1) will still test an
allocation requiring all pages, since the pages are always a
power of 2.